requirements.txt
opencv-python
pyserial
ultralytics